// src/app/Shared/Model/BookingInvoice.model.ts
export class BookingInvoiceModel {
    company: string;
    bookingcode: string;
    customer: string;
    sku: string;
    tourname: string;
    adults: number;
    children: number;
    created: string;
    date: string;
    departureTime: string;
    infants: number;
    orgdatetime: string;
    price: number;
    status: string;
    time: string;
}