// src/app/Shared/Model/MainBooking.model.ts
export class MainBookingModel {
    company: string;
    operator: string;
    bookcode: string;
    tourcode: string;
    tourname: string;
    customer: string;
    dateSubmitted: string;
    amount: number;
    noadult: number;
    noinfant: number;
    nochildren: number;
    tourdate: number;
    tourtime: string;
    customerphonenumber: string;
    created: string;
    bookingsource: string;
    authorizedbyAPI: string;
}