// src/app/Shared/Model/NewBooking.model.ts
export class NewBookingModel {
    company: string;
    bookingcode: string;
    customer: string;
    bookdate: string;
    bookingstatus: string;
    cfirstname: string;
    clastname: string;
    cphonenumber: string;
    created: string;
    date: string;
    modified: string;
    orgdatetime: string;
    paymentstatus: string;
    physicaladdress: string;
    time: string;
    totpersons: number;
    totprice: number;
    travelerType: string;
    tourcode: string;
    tourname: string;
}